спринт №21 отправка на ревью

+++ app/src/main/java/com/practicum/playlistmaker/db/FavoriteTrackEntity.kt
@@ -15,5 +15,5 @@ data class FavoriteTrackEntity(
     val country: String,
     val trackTimeMillis: Int,
     val previewUrl: String,
-    val addTimeStamp: Int = 0
+    val addTimeStamp: Long = 0
 )
